use relative path to load framework rules
We uses a path for the framework rules that was relative to the working
directory and not the loading file. This will lead to issues when using
a binary that will be located just somewhere.

This patch changes the loading path of the rule set file to be relative
to the FrameworkDetector.